feat: keyboard shortcuts ui gestures (#149)
Refactor the overlay-play component and rename it to overlay-action which can display overlay action gesture of any kind of icon.
Added store for the overlay-action component. Using this store the keyboard component (or any other component) can call updateOverlayActionIcon(...) and supply an icon/s. The overlay-action component will know to display the correct icon based on the updated iconType prop.
Fix the overlay animation to look like in Hillel's demo.