fix(FEC-11284): empty top bar fix (#610)
Moved back Margins to left / right containers instead of padding in the top bar
Removed padding in bottom bar when not shown and added delay in animation for it not to be visible while sliding out / in
Replaced traslate -y in top bar with margin-top because it didn't effect the interactive area
Reduced top bar and bottom bar appearance / disappearing animation to 100 to match the captions animation. If we to increase it we will need to increase them both.