fix(FEC-12372): [WEB][UI] - [O2CZ] Autoplay_WEB - After seeking to the end of the program, the next program will not start playing automatically
When a user performs a mousedown on the seekbar, and a mouseup on the overlay, and no mousedown event has been fired on the overlay earlier, overlay-action does not detect a drag. This causes the mousedown to toggle the player's play/pause.
The solution was to update the drag detection in isDragging by adding the seekbar drag state.